refactor: 파일명을 컨벤션에 맞게 kebab-case로 변경
